refactor: web3-telemetry
