+++ packages/mask/background/tsconfig.json
@@ -17,6 +17,7 @@
     { "path": "../../backup-format" },
     { "path": "../../gun-utils" },
     { "path": "../../flags" },
+    { "path": "../../web3-telemetry" },
     { "path": "../../web3-providers" },
     { "path": "../../sandboxed-plugin-runtime/src/background" }
   ]
