+++ packages/mask/src/setup.ui.0.ts
@@ -1,4 +1,5 @@
-import { TelemetryID, type KVStorageBackend } from '@masknet/shared-base'
+import { type KVStorageBackend } from '@masknet/shared-base'
+import { TelemetryID } from '@masknet/web3-telemetry'
 import { setupMaskKVStorageBackend } from '../shared/kv-storage.js'
 import { setupLegacySettingsAtNonBackground } from '../shared/legacy-settings/createSettings.js'
 import Services from './extension/service.js'
@@ -7,6 +8,7 @@ import { MaskMessages } from '../shared/messages.js'
 
 Services.Helper.getTelemetryID().then((id) => (TelemetryID.value = id))
 MaskMessages.events.telemetryIDReset.on((id) => (TelemetryID.value = id))
+
 const memory: KVStorageBackend = {
     beforeAutoSync: Promise.resolve(),
     getValue(...args) {
