+++ packages/shared-base/src/utils/index.ts
@@ -1,9 +1,7 @@
-export * from './ABTest.js'
 export * from './attachRectangle.js'
 export * from './ValueRef.js'
 export * from './asyncIterator.js'
 export * from './compose.js'
-export * from './TelemetryID.js'
 export * from './createLookupTableResolver.js'
 export * from './getLocalImplementation.js'
 export * from './markdown.js'
