+++ packages/web3-telemetry/src/entry.ts
@@ -1,0 +1,1 @@
+export * from './constants/index.js'
