+++ packages/web3-telemetry/tests/helpers/getABTestSeed.ts
@@ -1,11 +1,7 @@
 import { describe, it, expect } from 'vitest'
-import { joinsABTest, getABTestSeed } from '../../src/utils/ABTest.js'
+import { getABTestSeed } from '../../src/helpers/getABTestSeed.js'
 
 describe('detect if an installation joins the A/B test', () => {
-    it('should calculate join A/B test correctly', () => {
-        expect(joinsABTest(100)).toBeTruthy()
-        expect(joinsABTest(0)).toBeFalsy()
-    })
     it('should create an A/B seed in the given range', () => {
         const seed = getABTestSeed(16)
